requirements.txt
fastapi
uvicorn
openai
python-dotenv
sqlalchemy
pypdf
jinja2
python-multipart
